Moved FilledVariables Structure
Moved FilledVaraibles structure and functions to their own files
